d2::filter::scaled_filter::filter_channel(): Revise bounds checking to avoid unsigned/signed type warnings.
darcs-hash:20071025041434-d1b56-458e8169f347a3cb15754dae7f96487667be6e94.gz

+++ d2/filter/scaled_filter.h
@@ -286,12 +286,12 @@ private:
 
 				if (imin[0] < 0)
 					imin[0] = 0;
-				if (imax[0] > im->height() - 1)
-					imax[0] = im->height() - 1;
+				if (imax[0] > (int) im->height() - 1)
+					imax[0] = (int) im->height() - 1;
 				if (imin[1] < 0)
 					imin[1] = 0;
-				if (imax[1] > im->width() - 1)
-					imax[1] = im->width() - 1;
+				if (imax[1] > (int) im->width() - 1)
+					imax[1] = (int) im->width() - 1;
 			}
 
 			/*
